+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 - ✨ Add support for compiling `if_test` for all gatesets and devices ([#805]) ([**@adnathanail**])
+- ✨ Add benchmark for dynamical GHZ state generation ([#816]) ([**@TomasVF**])
 - ✨ Add Shor's 9 Qubit Code benchmark ([#803]) ([**@adnathanail**])
 - ✨ Add 7 Qubit Steane Code benchmark ([#814]) ([**@adnathanail**])
 
@@ -100,6 +101,7 @@ _📚 Refer to the [GitHub Release Notes] for previous changelogs._
 
 <!-- PR links -->
 
+[#816]: https://github.com/munich-quantum-toolkit/bench/pull/816
 [#814]: https://github.com/munich-quantum-toolkit/bench/pull/814
 [#805]: https://github.com/munich-quantum-toolkit/bench/pull/805
 [#803]: https://github.com/munich-quantum-toolkit/bench/pull/803
@@ -149,6 +151,7 @@ _📚 Refer to the [GitHub Release Notes] for previous changelogs._
 [**@denialhaag**]: https://github.com/denialhaag
 [**@soroushfathi**]: https://github.com/soroushfathi
 [**@adnathanail**]: https://github.com/adnathanail
+[**@TomasVF**]: https://github.com/TomasVF
 
 <!-- General links -->